Dupe encoding (#3261)
* Fix weird duplicate encoding issue

Sometimes (maybe when the server is under load and not responsive)
the client sends multiple copies of a file to the server, and diff
pods all try to encode the file at the same time, which both
is an inefficient use of resources. The errors from these also makes
it difficult to diagnose real FFmpeg issues from weird collision ones.
This PR checks if a file exists on S3 before beginning the encoding
process, and if yes, skips it. It incidentally also:

* Fixes issues where throw errors inside a request don't terminate
  the rest of the function
* Ensures tasks for after S3 upload actually happens after the entire
  stream is encoded

* Send encoding errors to Sentry instead of just logging

+++ server/src/lib/clip.ts
@@ -8,6 +8,7 @@ import Model from './model';
 import getLeaderboard from './model/leaderboard';
 import { earnBonus, hasEarnedBonus } from './model/achievements';
 import * as Basket from './basket';
+import * as Sentry from '@sentry/node';
 import Bucket from './bucket';
 import { ClientParameterError, ServerError } from './utility';
 import Awards from './model/awards';
@@ -69,6 +70,17 @@ export default class Clip {
     return router;
   }
 
+
+  /*
+   * Helper function to send error message to client, and save to Sentry
+   * defaults to save_clip_error
+   */
+  clipSaveError(headers: any, response: Response, status: number, msg: string, customError?: string) {
+    const compiledError = customError ? `${customError} ${msg}` : `save_clip_error ${msg}`;
+    response.status(status).send(compiledError);
+    Sentry.captureEvent({ request: { headers }, message: compiledError });
+  }
+
   serveClip = async ({ params }: Request, response: Response) => {
     const url = await this.bucket.getClipUrl(params.clip_id);
     if (url) {
@@ -79,25 +91,21 @@ export default class Clip {
   };
 
   saveClipVote = async (
-    { client_id, body, params }: Request,
+    { client_id, body, params, headers }: Request,
     response: Response
   ) => {
     const id = params.clipId as string;
     const { isValid, challenge } = body;
 
     if (!id || !client_id) {
-      response.statusMessage = 'save_clip_vote_missing_parameter';
-      response
-        .status(400)
-        .send(`Missing parameter: ${id ? 'client_id' : 'clip_id'}.`);
-      throw new ClientParameterError();
+      this.clipSaveError(headers, response, 400, `missing parameter: ${id ? 'client_id' : 'clip_id'}`, 'save_vote_error');
+      return;
     }
 
     const clip = await this.model.db.findClip(id);
     if (!clip) {
-      response.statusMessage = 'save_clip_vote_missing_clip';
-      response.status(422).send(`Clip not found: ${id}.`);
-      throw new ServerError();
+      this.clipSaveError(headers, response, 422, `clip not found: ${id}`, 'save_vote_error');
+      return;
     }
 
     const glob = clip.path.replace('.mp3', '');
@@ -134,7 +142,8 @@ export default class Clip {
 
   /**
    * Save the request body as an audio file.
-   * TODO: Check for empty or silent clips before uploading.
+   * Ensure errors from this function include the term save_clip_error
+   * to be easily parsed from other errors
    */
   saveClip = async (request: Request, response: Response) => {
     const { client_id, headers } = request;
@@ -144,29 +153,23 @@ export default class Clip {
     const size = headers['content-length'];
 
     if (!sentenceId || !client_id) {
-      response.statusMessage = 'save_clip_missing_parameter';
-      response
-        .status(400)
-        .send(
-          `Missing parameter: ${sentenceId ? 'client_id' : 'sentence_id'}.`
-        );
-      console.log(`sent headers: ${JSON.stringify(headers)}`);
-      throw new ClientParameterError();
+      this.clipSaveError(headers, response, 400, `missing parameter: ${sentenceId ? 'client_id' : 'sentence_id'}`);
+      return;
     }
 
     const sentence = await this.model.db.findSentence(sentenceId);
     if (!sentence) {
-      response.statusMessage = 'save_clip_missing_sentence';
-      response.status(422).send(`Sentence not found: ${sentenceId}.`);
-      throw new ServerError();
+      this.clipSaveError(headers, response, 422, `sentence not found: ${sentenceId}`);
+      return;
     }
 
     // Where is our audio clip going to be located?
     const folder = client_id + '/';
     const filePrefix = sentenceId;
     const clipFileName = folder + filePrefix + '.mp3';
+    const metadata = `${clipFileName} (${size} bytes, ${format}) from ${source}`;
 
-    try {
+    if (!await this.bucket.fileExists(clipFileName)) {
       // If the folder does not exist, we create it.
       await this.s3
         .putObject({ Bucket: getConfig().CLIP_BUCKET_NAME, Key: folder })
@@ -193,6 +196,51 @@ export default class Clip {
         .format('mp3')
         .channels(1)
         .sampleRate(32000)
+        .on('error', (error: string) => {
+          this.clipSaveError(headers, response, 500, `${error} for ${metadata}`);
+          return;
+        })
+        .on('finish', async() => {
+          console.log(`clip written to s3 ${metadata}`);
+
+          await this.model.saveClip({
+            client_id: client_id,
+            localeId: sentence.locale_id,
+            original_sentence_id: sentenceId,
+            path: clipFileName,
+            sentence: sentence.text,
+          });
+          await Awards.checkProgress(client_id, { id: sentence.locale_id });
+
+          await checkGoalsAfterContribution(client_id, { id: sentence.locale_id });
+
+          Basket.sync(client_id).catch(e => console.error(e));
+
+          const challenge = headers.challenge as ChallengeToken;
+          const ret = challengeTokens.includes(challenge)
+            ? {
+                filePrefix: filePrefix,
+                showFirstContributionToast: await earnBonus('first_contribution', [
+                  challenge,
+                  client_id,
+                ]),
+                hasEarnedSessionToast: await hasEarnedBonus(
+                  'invite_contribute_same_session',
+                  client_id,
+                  challenge
+                ),
+                // can't simply reduce the number of the calls to DB through streak_days in checkGoalsAfterContribution()
+                // since the the streak_days may start before the time when user set custom_goals, check to win bonus for each contribution
+                showFirstStreakToast: await earnBonus('three_day_streak', [
+                  client_id,
+                  client_id,
+                  challenge,
+                ]),
+                challengeEnded: await this.model.db.hasChallengeEnded(challenge),
+              }
+            : { filePrefix };
+          response.json(ret);
+        })
         .stream();
 
       await this.s3
@@ -202,53 +250,9 @@ export default class Clip {
           Body: audioOutput,
         })
         .promise();
-
-      console.log(
-        `clip written to s3 ${clipFileName} (${size} bytes, ${format}) from ${source}`
-      );
-
-      await this.model.saveClip({
-        client_id: client_id,
-        localeId: sentence.locale_id,
-        original_sentence_id: sentenceId,
-        path: clipFileName,
-        sentence: sentence.text,
-      });
-      await Awards.checkProgress(client_id, { id: sentence.locale_id });
-
-      await checkGoalsAfterContribution(client_id, { id: sentence.locale_id });
-
-      Basket.sync(client_id).catch(e => console.error(e));
-
-      const challenge = headers.challenge as ChallengeToken;
-      const ret = challengeTokens.includes(challenge)
-        ? {
-            filePrefix: filePrefix,
-            showFirstContributionToast: await earnBonus('first_contribution', [
-              challenge,
-              client_id,
-            ]),
-            hasEarnedSessionToast: await hasEarnedBonus(
-              'invite_contribute_same_session',
-              client_id,
-              challenge
-            ),
-            // can't simply reduce the number of the calls to DB through streak_days in checkGoalsAfterContribution()
-            // since the the streak_days may start before the time when user set custom_goals, check to win bonus for each contribution
-            showFirstStreakToast: await earnBonus('three_day_streak', [
-              client_id,
-              client_id,
-              challenge,
-            ]),
-            challengeEnded: await this.model.db.hasChallengeEnded(challenge),
-          }
-        : { filePrefix };
-      response.json(ret);
-    } catch (error) {
-      console.error(error);
-      response.statusCode = error.statusCode || 500;
-      response.statusMessage = 'save_clip_error';
-      response.json(error);
+    } else {
+      this.clipSaveError(headers, response, 500, `${clipFileName} already exists`);
+      return;
     }
   };
